v ext: code review 2
